requirements.txt
# Django y DRF
Django==5.2.7
djangorestframework==3.15.2

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# Filtros para DRF
django-filter==24.3

# Base de datos
psycopg[binary]==3.2.3

# CORS para conectar con frontend
django-cors-headers==4.6.0

# Variables de entorno
python-dotenv==1.0.1

# Para parsing de URLs de base de datos
dj-database-url==2.3.0

# Para validaciones
Pillow==11.0.0

# Para producción
gunicorn==23.0.0

# Para testing
pytest==8.3.4
pytest-django==4.9.0